implemented xgboost model
