+++ data_files/result/time_to_failure.csv
@@ -1,0 +1,17 @@
+0
+1.4307971859
+1.3914988931
+1.3531960947
+1.3137978019
+1.2743995091
+1.2360967106999998
+1.1967984179
+1.1584956195
+1.1190973267
+1.0796990339
+1.0413962355
+1.0019979427
+0.96269964992
+0.92439685152
+0.8849985587200001
+0.84669576032
